requirement.txt
accelerate==1.11.0
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.4.0
autoviz==0.1.905
backports.tarfile==1.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
bokeh==3.8.0
cachetools==6.2.1
category_encoders==2.7.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.0
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.6
dash==3.2.0
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
docutils==0.22.2
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
holoviews==1.21.0
huggingface-hub==0.36.0
hvplot==0.12.1
id==1.5.0
idna==3.11
imbalanced-learn==0.14.0
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_widgets==3.0.15
kaleido==1.1.0
keyring==25.6.0
kiwisolver==1.4.9
lightgbm==4.6.0
linkify-it-py==2.0.3
llvmlite==0.45.1
logistro==2.0.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplot==0.1.9
matplotlib==3.7.5
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.2
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.4
packaging==25.0
pandas==2.1.4
pandas-dq==1.29
panel==1.8.2
param==2.2.1
parso==0.8.5
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.0
plotly==6.3.1
plotly-resampler==0.11.0
pluggy==1.6.0
pmdarima==2.0.4
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.2
pure_eval==0.2.3
pyamg==5.3.0
pyarrow==21.0.0
pycaret==3.3.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyloco==0.0.139
pyod==2.0.5
pyparsing==3.2.5
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyviz_comms==3.0.6
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
readme_renderer==44.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
rfc3986==2.0.0
rich==14.2.0
rpds-py==0.28.0
safetensors==0.6.2
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.43.0
shap==0.49.1
simplejson==3.20.2
SimpleWebSocketServer==0.1.2
six==1.17.0
sktime==0.26.0
slicer==0.0.8
smmap==5.0.2
soundfile==0.13.1
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
streamlit==1.51.0
sympy==1.14.0
tbats==1.1.3
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
tsdownsample==0.1.4.1
twine==6.2.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
ushlex==0.99.1
wandb==0.22.3
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
xgboost==1.6.2
xlrd==2.0.2
xxhash==3.6.0
xyzservices==2025.10.0
yellowbrick==1.5
zipp==3.23.0